large amount of cleanup and refactoring
moved most of SWF services to shared services as they're used in multiple areas.  Wired up DWF forecast queries.  Now returns valid weather using recent locations and new user inputs.
