+++ src/services/DWFServices.js
@@ -1,5 +1,4 @@
 import firebase from "../firebaseConfig";
-import {zipToGeo} from '@/services/SWFServices'
 const db = firebase;
 
 export async function getAllRecentLocations(count) {
@@ -15,25 +14,4 @@ export async function getAllRecentLocations(count) {
   })
 
   return locArray
-}
-
-
-export async function checkDbFor(zip) {
-  const docRef = db.collection('geo').doc(zip)
-
-  const geoData = await docRef.get()
-  console.log('geoData', geoData)
-  //
-  // return docRef.get()
-  //   .then(async (doc) => {
-  //     if (doc.exists)
-  //       return doc.data()
-  //     else {
-  //       let geoData = await zipToGeo(zip)
-  //       // TODO: fix geoData and add state logic
-  //       docRef.set(geoData)
-  //       return geoData
-  //     }
-  //   })
-
 }
